backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
mediapipe==0.10.8
# tensorflow==2.14.0
# numpy==1.24.3
# deepface==0.0.79  # Removed - using custom CNN model instead
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
pymongo==4.6.1
pydantic-settings==2.1.0
bcrypt==4.0.1
email-validator==2.1.0
python-slugify==8.0.1
slowapi==0.1.8
redis==5.0.1
psutil==5.9.8
requests==2.31.0